requirements.txt
elevenlabs==1.13.0
exceptiongroup==1.2.0
executing==0.8.3
GitPython==3.1.43
google-api-core==2.19.2
google-api-python-client==2.154.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-vision==3.7.4
google_search_results==2.4.2
googleapis-common-protos==1.65.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.8
langchain-experimental==0.3.2
langchain-openai==0.2.1
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.130
layoutparser==0.3.4
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
pip==24.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
python-poppler==0.4.1
pytz==2024.1
pyvis==0.3.2
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rapidfuzz==3.9.7
referencing==0.30.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0

